We now concatenate the head and modifier hidden states for dual mode classification.
