+++ encoder/src/main/python/trainer.py
@@ -6,7 +6,7 @@ import pandas as pd
 
 from torch.utils.data import Dataset
 
-from transformers import AutoTokenizer, AutoConfig, TrainingArguments, Trainer, DataCollatorForTokenClassification
+from transformers import AutoTokenizer, AutoConfig, TrainingArguments, Trainer
 from datasets import Dataset, DatasetDict
 
 import time
@@ -15,60 +15,8 @@ from datetime import timedelta
 import configuration as cf
 from task import Task
 from token_classifier import TokenClassificationModel
+from dual_data_collator import OurDataCollator
 from evaluation_metrics import evaluation_classification_report, evaluate
-
-class OurDataCollator(DataCollatorForTokenClassification):
-  def make_head_features(self, features):
-    head_feats = []
-    for feature in features:
-      head_dict = {}
-      head_dict['input_ids'] = feature[cf.HEAD_POSITIONS]
-      head_feats.append(head_dict)
-    return head_feats
-
-  def torch_call(self, features):
-        import torch
-
-        label_name = "label" if "label" in features[0].keys() else "labels"
-        labels = [feature[label_name] for feature in features] if label_name in features[0].keys() else None
-
-        batch = self.tokenizer.pad(
-            features,
-            padding=self.padding,
-            max_length=self.max_length,
-            pad_to_multiple_of=self.pad_to_multiple_of,
-            # Conversion to tensors will fail if we have labels as they are not of the same length yet.
-            return_tensors="pt" if labels is None else None,
-        )
-        # HF does not batch head_positions, so we have to fake it by masquerading them as input_ids
-        heads = self.make_head_features(features)
-        batch_heads = self.tokenizer.pad(
-            heads,
-            padding=self.padding,
-            max_length=self.max_length,
-            pad_to_multiple_of=self.pad_to_multiple_of,
-            # Conversion to tensors will fail if we have labels as they are not of the same length yet.
-            return_tensors="pt" if labels is None else None,
-        )
-        batch[cf.HEAD_POSITIONS] = batch_heads['input_ids']
-
-        if labels is None:
-            return batch
-
-        sequence_length = torch.tensor(batch["input_ids"]).shape[1]
-        padding_side = self.tokenizer.padding_side
-        if padding_side == "right":
-            batch[label_name] = [
-                list(label) + [self.label_pad_token_id] * (sequence_length - len(label)) for label in labels
-            ]
-        else:
-            batch[label_name] = [
-                [self.label_pad_token_id] * (sequence_length - len(label)) + list(label) for label in labels
-            ]
-
-        batch = {k: torch.tensor(v, dtype=torch.int64) for k, v in batch.items()}
-
-        return batch
 
 # main function for training the MTL classifier
 def main():
